edits

+++ FR2017/Digits.java
@@ -11,8 +11,8 @@ public class Digits{
     }
     public boolean isStrictlyIncreasing(){
         for (int i=1; i<digitList.size(); i++){
-            if (digitList.get(i-1).intValue()>=digitList.get(i).intValue()) return false;
-        }
+            if (digitList.get(i-1)>=digitList.get(i)) return false; // auto unboxing Integer() to int
+        }                                                           // could use intValue() or equals()
         return true;
     }
 }
